requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Database connectivity
psycopg2-binary>=2.9.0  # For Redshift/PostgreSQL connections

# Data validation and parsing
jsonschema>=4.0.0

# Logging and configuration
pyyaml>=6.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: For enhanced data processing
# sqlalchemy>=1.4.0
# boto3>=1.26.0  # For S3 integration if needed 